Consolidate PDF generation integration tests
The empty/non-item publish and preview guards in pdf_generation_test
and delivery_note_pdf_test duplicated publishable_document_test, which
exercises the shared PublishableDocument/DocumentWithLines filters via
the Invoice routes. Dropped those copies.

One exception: delivery_notes_controller#publish has its own inline
publish_problems guard (not the shared concern, unlike the preview
filter), so its empty-publish test is kept — verified via SimpleCov
that removing it dropped coverage of controller lines 96-100.

invoice_language_test's EN/DE/switching tests only asserted the bytes
start with "%PDF" — pure FOP smoke already covered here and in the
renderer tests. Removed them and the two customer-language tests
(duplicated in customer_test); moved the genuine export/VAT-id
regression into this file and deleted invoice_language_test. Promoted
the triplicated assert_valid_pdf_response into a shared helper.
